django_app/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
requests>=2.31
python-dotenv>=1.0
